Move to Twig package, add the Twig runtime extension interface on the runtime class
