+++ composer.json
@@ -12,6 +12,7 @@
     "require-dev": {
         "doctrine/collections": "^1.4",
         "doctrine/dbal": "^2.5",
+        "doctrine/mongodb-odm": "^2.0",
         "doctrine/orm": "^2.5",
         "doctrine/phpcr-odm": "^1.3",
         "friendsofphp/php-cs-fixer": "^2.16.3",
@@ -24,10 +25,10 @@
         "propel/propel1": "^1.7",
         "ruflin/elastica": "^1.3 || ^2.0 || ^3.0 || ^5.0 || ^6.0 || ^7.0",
         "solarium/solarium": "^2.3 || ^3.0 || ^4.0 || ^5.0",
-        "twig/twig": "^1.34 || ^2.4 || ^3.0"
+        "twig/twig": "^1.35 || ^2.5 || ^3.0"
     },
     "conflict": {
-        "twig/twig": "<1.34 || >=2.0,<2.4"
+        "twig/twig": "<1.35 || >=2.0,<2.5"
     },
     "suggest": {
         "twig/twig": "To integrate Pagerfanta with Twig"
@@ -40,7 +41,8 @@
         "pagerfanta/doctrine-orm-adapter": "self.version",
         "pagerfanta/doctrine-phpcr-odm-adapter": "self.version",
         "pagerfanta/elastica-adapter": "self.version",
-        "pagerfanta/solarium-adapter": "self.version"
+        "pagerfanta/solarium-adapter": "self.version",
+        "pagerfanta/twig": "self.version"
     },
     "autoload": {
         "psr-4": {
@@ -52,7 +54,8 @@
             "Pagerfanta\\Doctrine\\ORM\\": "lib/Adapter/Doctrine/ORM/",
             "Pagerfanta\\Doctrine\\PHPCRODM\\": "lib/Adapter/Doctrine/PHPCRODM/",
             "Pagerfanta\\Elastica\\": "lib/Adapter/Elastica/",
-            "Pagerfanta\\Solarium\\": "lib/Adapter/Solarium/"
+            "Pagerfanta\\Solarium\\": "lib/Adapter/Solarium/",
+            "Pagerfanta\\Twig\\": "lib/Twig/"
         }
     },
     "autoload-dev": {
