+++ lib/Core/RouteGenerator/RouteGeneratorDecorator.php
