+++ lib/Twig/composer.json
@@ -1,0 +1,22 @@
+{
+    "name": "pagerfanta/twig",
+    "type": "library",
+    "description": "Twig integration for Pagerfanta",
+    "keywords": ["pagerfanta"],
+    "license": "MIT",
+    "require": {
+        "php": "^7.2",
+        "twig/twig": "^1.35 || ^2.5 || ^3.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "Pagerfanta\\Twig\\": "./"
+        }
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-2.x": "2.x-dev"
+        }
+    },
+    "minimum-stability": "dev"
+}
